bds: Bug fixes in carrier wavelength computation
Changed wavelength computation to native method within RTKLIb. The native method
 should also help in the PVT computation of GLONASS satellites as it considers
the freq. offset when computing the wavelength. This also fixes typo in tracking
 stage for MEO satellites of the BDS constellation.

Note: A similar approach to carrier wavelength computation should be applied to
the obs_data in rtk_solver. In theory, static allocation should not be needed for
the types.